health: set state to WaitingToRegenerate in LaunchAsEndpoint
We need to ensure that the health daemon successfully regenerates synchronously
when LaunchAsEndpoint is executed. So, we need to ensure that the health
endpoint is in a valid state to regenerate, otherwise the controller which
launches the health endpoint will always fail beacause the health endpoint will
never transition out of 'ready' state which is not a valid state from which it
can regenerate.

Signed-off by: Ian Vernon <ian@cilium.io>